Refactor: Remove source/source_pw/module_ofdft/ml_tools
