+++ source/source_estate/module_pot/pot_ml_exx.h
@@ -5,7 +5,7 @@
 #ifdef __MLALGO
 
 #include "pot_base.h"
-#include "source_pw/module_ofdft/ml_tools/nn_of.h"
+#include "source_pw/module_ofdft/nn_of.h"
 #include "source_io/module_ml/cal_mlkedf_descriptors.h"
 #include "source_pw/module_ofdft/ml_base.h"
