+++ source/source_pw/module_ofdft/CMakeLists.txt
@@ -26,7 +26,7 @@ if(ENABLE_MLALGO)
       kedf_ml_label.cpp
       ml_base.cpp
       ml_base_pot.cpp
-      ml_tools/nn_of.cpp
+      nn_of.cpp
   )
 
   add_library(
