+++ source/source_pw/module_ofdft/ml_base.h
@@ -8,7 +8,7 @@
 #include <cmath>
 
 #ifdef __MLALGO
-#include "source_pw/module_ofdft/ml_tools/nn_of.h"
+#include "source_pw/module_ofdft/nn_of.h"
 #include "source_io/module_ml/cal_mlkedf_descriptors.h"
 
 // The ML_Base class encapsulates common functionality for Machine Learning based
